requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
ansicolor=0.2.6=pypi_0
appdirs=1.4.4=pypi_0
appnope=0.1.0=pypi_0
asgiref=3.2.10=py_0
astor=0.8.1=pypi_0
astroid=2.4.2=py38_0
attrs=20.2.0=py_0
autoflake=1.4=pypi_0
autopep8=1.5.4=py_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
ca-certificates=2020.7.22=0
certifi=2020.6.20=py38_0
chardet=3.0.4=pypi_0
colorama=0.4.3=pypi_0
decorator=4.4.2=pypi_0
distlib=0.3.1=pypi_0
django=3.1.1=py_0
django-extensions=3.0.9=pypi_0
django-filter=2.3.0=py_0
django-model-utils=4.0.0=py_0
django-pandas=0.6.1=py_0
django-redis=4.12.1=pypi_0
django-webpack-loader=0.7.0=pyh8c360ce_0
djangorestframework=3.11.1=py_0
epc=0.0.5=py_0
filelock=3.0.12=pypi_0
flake8=3.8.3=py_0
funcparserlib=0.3.6=pypi_0
future=0.18.2=py38_1
greenlet=0.4.17=pypi_0
hy=0.19.0=pypi_0
importlib-metadata=1.7.0=py38_0
importmagic=0.1.7=pypi_0
iniconfig=1.0.1=py_0
intel-openmp=2019.4=233
ipdb=0.13.3=pypi_0
ipython=7.18.1=pypi_0
ipython-genutils=0.2.0=pypi_0
isort=5.5.3=py38_0
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
json-rpc=1.13.0=pypi_0
krb5=1.17.1=hddcf347_0
lazy-object-proxy=1.4.3=py38h1de35cc_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libpq=12.2=h051b688_0
markdown=3.2.2=py38_0
markupsafe=1.1.1=py38h1de35cc_1
mccabe=0.6.1=pypi_0
mkl=2019.4=233
mkl-service=2.3.0=py38hfbe908c_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
more-itertools=8.5.0=py_0
msgpack=1.0.0=pypi_0
mypy=0.782=pypi_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.2=h0a44026_1
numpy=1.19.1=py38h3b9f5b6_0
numpy-base=1.19.1=py38hcfb5961_0
openssl=1.1.1h=haf1e3a3_0
packaging=20.4=py_0
pandas=1.1.1=py38hb1e8313_0
parso=0.7.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pip=20.2.3=pypi_0
pluggy=0.13.1=py38_0
prompt-toolkit=3.0.7=pypi_0
psycopg2=2.8.5=py38hddc9c9b_0
ptvsd=4.3.2=pypi_0
ptyprocess=0.6.0=pypi_0
py=1.9.0=py_0
pycodestyle=2.6.0=py_0
pydocstyle=5.1.1=py_0
pyflakes=2.2.0=py_0
pygments=2.7.1=pypi_0
pylint=2.6.0=py38_0
pyls-isort=0.2.0=pypi_0
pyls-mypy=0.1.8=pypi_0
pynvim=0.4.2=pypi_0
pyparsing=2.4.7=py_0
pytest=6.0.2=py38_0
pytest-django=3.10.0=pyh9f0ad1d_0
pytest-mock=3.2.0=py_0
python=3.8.5=h26836e1_1
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.35.1=py_0
pytz=2020.1=py_0
pyyaml=5.3.1=pypi_0
readline=8.0=h1de35cc_0
redis=3.5.3=pypi_0
rope=0.17.0=py_0
rply=0.7.7=pypi_0
service-factory=0.1.6=pypi_0
setuptools=49.6.0=py38_0
sexpdata=0.0.3=pypi_0
six=1.15.0=py_0
snowballstemmer=2.0.0=py_0
solargraph-utils-py=1.1.0=pypi_0
sqlite=3.33.0=hffcf06c_0
sqlparse=0.3.1=py_0
tk=8.6.10=hb0a8c7a_0
toml=0.10.1=py_0
traitlets=5.0.4=pypi_0
typed-ast=1.4.1=pypi_0
typing-extensions=3.7.4.3=pypi_0
ujson=3.2.0=py38hb1e8313_0
vim-vint=0.3.21=pypi_0
virtualenv=20.0.31=pypi_0
wcwidth=0.2.5=pypi_0
wheel=0.35.1=py_0
wrapt=1.11.2=py38h1de35cc_0
xz=5.2.5=h1de35cc_0
yapf=0.30.0=py_0
zipp=3.1.0=py_0
zlib=1.2.11=h1de35cc_3
